+++ CHANGELOG.md
@@ -7,6 +7,7 @@ GAMSPy 0.14.2
   - Add generate_name_dict option.
   - Disable solution report by default.
   - Fix the order of equations in toGams utility.
+  - Allow options in toGams.
 
 GAMSPy 0.14.1
 -------------